Break up character importer
Split out character import into smaller files
