+++ src/parser/character/currency.js
@@ -1,0 +1,9 @@
+export function getCurrency(data) {
+  return {
+    pp: data.character.currencies.pp,
+    gp: data.character.currencies.gp,
+    ep: data.character.currencies.ep,
+    sp: data.character.currencies.sp,
+    cp: data.character.currencies.cp,
+  };
+}
