+++ src/parser/index.js
@@ -1,9 +1,9 @@
 import getCharacter from "./character/index.js";
-import getFeatures from "./character/features.js";
-import getClasses from "./character/classes.js";
-import { default as getSpells, parseItemSpells as getItemSpells } from "./character/spells.js";
+import getActions from "./character/features/actions.js";
+import getFeatures from "./character/features/features.js";
+import getClasses from "./character/classes/index.js";
+import { default as getSpells, parseItemSpells as getItemSpells } from "./spells/index.js";
 import getInventory from "./inventory/index.js";
-import getActions from "./character/actions.js";
 
 let parseJson = (ddb) => {
   let character = getCharacter(ddb);
